Add TAD and aggregate TFT
